Move utility functions to respective files in util/
