+++ src/util/cleanObject.ts
@@ -1,0 +1,7 @@
+export function cleanObject(object: Object): void {
+  for (const key in object) {
+    if (object[key] == null) {
+      delete object[key]
+    }
+  }
+}
