+++ src/util/pinoLogger.ts
@@ -1,0 +1,5 @@
+import pino from 'pino'
+
+export const pinoLogger = pino({
+  level: process.env.VERBOSE == null ? 'info' : 'debug'
+})
