+++ src/util/snooze.ts
@@ -1,0 +1,2 @@
+export const snooze: Function = async (ms: number) =>
+  new Promise((resolve: Function) => setTimeout(resolve, ms))
